+++ tests/NamespaceTest.php
@@ -6,7 +6,6 @@ use PhpEditor\File;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @covers \PhpEditor\File
  * @covers \PhpEditor\Namespace_
  */
 class NamespaceTest extends TestCase
@@ -14,33 +13,33 @@ class NamespaceTest extends TestCase
     public function testGetNamespace()
     {
         $file = File::createFromSource('<?php namespace Foo;');
-        $this->assertEquals('Foo', $file->getNamespace());
+        $this->assertEquals('Foo', $file->getNamespace()->getName());
     }
 
     public function testGetNamespaceFromEmptyFile()
     {
         $file = File::createFromSource('<?php ');
-        $this->assertNull($file->getNamespace());
+        $this->assertNull($file->getNamespace()->getName());
     }
 
     public function testSetNamespace()
     {
         $file = File::createFromSource('<?php namespace Foo;');
-        $file->setNamespace('Bar\Baz');
+        $file->getNamespace()->setName('Bar\Baz');
         $this->assertEquals('<?php namespace Bar\Baz;', $file->getSource());
     }
 
     public function testSetNamespaceEmptyFile()
     {
         $file = File::create();
-        $file->setNamespace('Bar\Baz');
+        $file->getNamespace()->setName('Bar\Baz');
         $this->assertEquals("<?php\n\nnamespace Bar\Baz;\n", $file->getSource());
     }
 
     public function testHasNamespace()
     {
         $file = File::createFromSource('<?php namespace Foo;');
-        $this->assertTrue($file->hasNamespace());
+        $this->assertTrue($file->getNamespace()->exists());
     }
 
     public function testNoMultipleNamespaces()
@@ -54,20 +53,6 @@ class NamespaceTest extends TestCase
     public function testHasNamespaceNotDefined()
     {
         $file = File::createFromSource('<?php ');
-        $this->assertFalse($file->hasNamespace());
-    }
-
-    public function testGetOrCreateOpening()
-    {
-        $file = File::createFromSource('<?php echo "Hello";');
-        $token = $file->getOrCreateOpening();
-        $this->assertEquals('<?php ', $token->getValue());
-    }
-
-    public function testGetOrCreateOpeningNotExisting()
-    {
-        $file = File::create();
-        $token = $file->getOrCreateOpening();
-        $this->assertEquals("<?php\n", $token->getValue());
+        $this->assertFalse($file->getNamespace()->exists());
     }
 }