+++ CHANGELOG.md
@@ -4,4 +4,5 @@ Changelog
 v0.1.0 (not released)
 ---------------------
 
-- Namespace manipulation
+- Use statements edition
+- Namespace edition